M3-1435: Add tags panel to domain details (#4059)
* M3-1435: move tags pannel to a separate component

* M3-1435: Add tags panel to domain details

* M3-1435: Clean code

* Update comments

* M3-1435: remove binds

* M3-1435: minor code tweaks
