+++ src/components/TagsPanel/index.ts
@@ -1,0 +1,1 @@
+export { default, Props as TagsPanelProps } from './TagsPanel';
